requirements.txt
aiogram>=3.23.0
aiofiles>=25.1.0
python-dotenv>=1.2.1
sqlalchemy>=2.0.45
alembic>=1.17.2
aiohttp>=3.13.2
cryptography>=46.0.3
pydantic>=2.12.5
pydantic-settings>=2.12.0
asyncpg>=0.30.0

